migration fiel

+++ api/models.py
@@ -3,4 +3,4 @@ from django.db import models
 # Create your models here.
 class PlayList(models.Model):
     id_user = models.IntegerField()
-    id_music_deezer = models.IntegerField()
+    id_music_deezer = models.BigIntegerField()
